moved reverse DNS lookup to utils
