+++ collector/interface.go
@@ -1,11 +1,9 @@
 package collector
 
 import (
-	"net"
 	"strings"
 	"time"
 
-	"github.com/patrickmn/go-cache"
 	"github.com/prometheus/client_golang/prometheus"
 )
 
@@ -73,19 +71,4 @@ type worker struct {
 
 func (w worker) podName() string {
 	return strings.Split(w.Worker.IP, ".")[0]
-}
-
-// ReverseDNSLookup returns hostname if no results are found
-func ReverseDNSLookup(host string, dnsCache *cache.Cache) string {
-	if cacheValue, found := dnsCache.Get(host); found {
-		return cacheValue.(string)
-	}
-	names, err := net.LookupAddr(host)
-	if err != nil || len(names) == 0 {
-		dnsCache.Set(host, host, cache.DefaultExpiration)
-		return host
-	}
-	dnsResult := strings.Split(names[0], ".")[0]
-	dnsCache.Set(host, dnsResult, cache.DefaultExpiration)
-	return dnsResult
 }
